FormatExceptionPolicy implementations need to implement Serializable and equals/hashCode with the same constraints as SerializableFileFilter.  Refactored the `toBytes()` method into the interface `NoLambda`, and moved `SerializableFileFilter.EqualityBasedOnSerialization` to `NoLambda.EqualityBasedOnSerialization`.  Technically a breaking change, but will affect exactly zero real-world client code.
