+++ lib/src/main/java/com/diffplug/spotless/FormatExceptionPolicy.java
@@ -20,7 +20,7 @@ import java.io.Serializable;
 import java.nio.file.Path;
 
 /** A policy for handling exceptions in the format. */
-public interface FormatExceptionPolicy extends Serializable {
+public interface FormatExceptionPolicy extends Serializable, NoLambda {
 	/** Called for every error in the formatter. */
 	void handleError(Throwable e, FormatterStep step, File file, Path rootDir);
